Fix 2FA auth
The authenticate method raises a ProtonAPIA2FANeeded exception when it
internally calls the refresh() method. After that happens, the caller
needs to catch the exception and **needs to know** that it still needs to
call the refresh() method, to refresh the VPN session info.

Due to this, I think it's better to clearly split authenticate() and
refresh(), so that things are more transparent for the caller.