Add workaround for invalid session pop-up when refreshing tokens [VPNLINUX-656]
Based on my investigation the tokens were being refreshed correctly for the first time, but they were never being stored to the keyring and the old ones were persisting. So on a next API call they both would be invaliadated and a new login was required.

I've implemented a new workaround this issue. It's not a total fix but a description of the issue and the consquences of doing what I did are described in the Session class.